+++ src/main/java/com/dexma/hometest/domain/Cash.java
@@ -1,9 +1,12 @@
 package com.dexma.hometest.domain;
 
+import java.math.BigDecimal;
+
+
 /**
  *
  */
 public interface Cash
 {
-    double getValue();
+    BigDecimal getValue();
 }
